+++ docs/App.vue
@@ -40,7 +40,7 @@
         <footer class="footer">
             <div class="container">
                 <div class="content">
-                    <p><strong>Buefy</strong> by <a href="https://twitter.com/rafaelpimpa" target="_blank">Rafael Beraldo</a>.</p>
+                    <p>Made with <b-icon icon="favorite" color="is-danger"></b-icon> by <a href="https://twitter.com/rafaelpimpa" target="_blank">Rafael Beraldo</a>.</p>
                     <p class="mb-0"><small>The source code is licensed <a href="http://opensource.org/licenses/mit-license.php" target="_blank">MIT</a>.</small></p>
                     <p><small>The website content is licensed <a href="http://creativecommons.org/licenses/by-nc-sa/4.0/" target="_blank">CC ANS 4.0</a>.</small></p>
                     <div class="field has-addons">
@@ -75,7 +75,7 @@
                 const height = 400
                 const left = (window.screen.width - width) / 2
                 const top = (window.screen.height - height) / 2
-                const url = `https://twitter.com/share?url=${window.location.href}&Buefy: a lightweight Bulma UI components for Vue.js&hashtags=buefy&via=rafaelpimpa`
+                const url = `https://twitter.com/share?url=${window.location.href}&Buefy: lightweight Bulma UI components for Vue.js&hashtags=buefy&via=rafaelpimpa`
                 const opts = `status=1,width=${width},height=${height},top=${top},left=${left}`
 
                 window.open(url, '', opts)