+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Buefy Changelog
 
+## 0.0.3
+
+* New component: Message
+* New component: Tooltip
+* Added size option to Pagination
+* Added color option to Icon
+* Notification slightly modified to keep original Bulma style
+
 ## 0.0.2
 
 * New component: Pagination